Modified LoopCollectionView size, added image view above it. Moved error message and corner radius enum to solo file.
