+++ Hypno Loops/ViewControllers/CreateAccountController.swift
@@ -21,29 +21,6 @@ class CreateAccountController: UIViewController {
     @IBOutlet weak var submitButtton: UIButton!
     @IBOutlet weak var skipButton: UIButton!
 
-    enum ErrorMessage {
-        case invalidUsername, invalidEmail, invalidPassword_Count,
-             invalidPassword_NeedDigit, invalidPassword_NeedUppercase,
-             invalidPassword_NeedLowercase
-        
-        var displayError: String {
-            switch self {
-            case .invalidUsername:
-                return "Username must have at least 3 characters"
-            case .invalidEmail:
-                return "Invalid Email Address"
-            case .invalidPassword_Count:
-                return "Password must be at least 8 characters"
-            case .invalidPassword_NeedDigit:
-                return "Password must contain at least 1 digit"
-            case .invalidPassword_NeedUppercase:
-                return "Password must contain at least 1 uppercase letter"
-            case .invalidPassword_NeedLowercase:
-                return "Password must contain at least 1 lowercase letter"
-            }
-        }
-    }
-    
     enum requiredText: String {
         case Required, Success
     }
@@ -51,17 +28,12 @@ class CreateAccountController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         resetForm()
-        
-        //usernameTextField.isHidden = true
-        //usernameErrorLabel.isHidden = true
+ 
         passwordTextField.isSecureTextEntry = true
     }
     
     func resetForm() {
-        //usernameErrorLabel.isHidden = false
-        //emailErrorLabel.isHidden = false
-        //passwordErrorLabel.isHidden = false
-       
+     
         usernameErrorLabel.text = requiredText.Required.rawValue
         emailErrorLabel.text = requiredText.Required.rawValue
         passwordErrorLabel.text = requiredText.Required.rawValue
@@ -248,7 +220,7 @@ class CreateAccountController: UIViewController {
     //MARK: - Segue
     
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        if segue.identifier == "gotoProfile" {
+        if segue.identifier == SegueID.gotoProfile.rawValue {
             let destinationVC = segue.destination as! UserProfileController
         }
     }
